Use pure="on" for Go binaries
This is equivalent to CGO_ENABLED=0 in the Go build system. It enhances
binary compatibility by avoiding a dependency on the host's libc.

In particular, if you build on a glibc 2.32 system, this avoids the
error:

/setup-robot-files/synk: /lib/x86_64-linux-gnu/libc.so.6: version
`GLIBC_2.32' not found

Setting pure="on" for synk was sufficient to fix that problem, but it
seems sensible to set it for all binaries. The only downside I could
find is that we might encounter surprising DNS resolution behavior,
because without cgo you can only use the "pure Go resolver":
https://pkg.go.dev/net#hdr-Name_Resolution

I also updated rules_docker in an attempt to resolve the issue, although
it didn't help.

Change-Id: I71ab96809a26f1b27225b4e8b6002e500f4cd7da
GitOrigin-RevId: 80dbab220102c026b555e07d3601ceba90e64aa6