Mirror fix for https://github.com/GoogleCloudPlatform/kubernetes-engine-customize-fluentd/issues/18 to our config.
I have observed this export error in our fluentd logs.

Note that this refactors the config so that the google-fluentd.conf coming from the docker image is overwritten on startup. igorp@ has confirmed that the current setup creates two slightly different competing output plugins. With this change, there should only be the single one we explicitly configure.

b/205263713

Change-Id: I2811a82fe41a74fda4b9f5630aff64626044afcb
GitOrigin-RevId: 5958b7584b3065bd3fb73fc3279a3430bf280b11