Change condition for Robot CR creation
Before, it depended on the value of 3 flags, which could be confusing as
it wasn't clear that setting any of them would lead to CR creation.

Now, it's set based on the --cr-syncer flag. While this is also a bit
indirect, it seems more natural, and saves the difficulty of adding a
new flag.

Tested by running setup-robot with/without --cr-syncer=false.

Change-Id: Icdbf6d69ce74b823fbe06243ec109f194d00da06
GitOrigin-RevId: 0dfb5f9a2d245d6411fd20201824acb456544cf5