setup_robot: bump --pod-running-timeout from 1m to 3m
If a single-node cluster has just been rebooted, it can take around a
minute from apiserver startup to setup-robot pod being able to start
(not sure exactly why, but I guess this is tied to CNI startup and
kubelet generally being a bit slow).

The default timeout of 60s is a little too close for comfort in this
case - if it takes any longer, you just see:

    pod "setup-robot" deleted
    error: timed out waiting for the condition

This should make that error case less likely.

Note: we saw that error consistently after reboot with Kubernetes
v1.20.4, but upgrading to v1.20.11 mostly avoids it by improving kubelet
startup.

b/203744154

Change-Id: I05dd3e0ff2600745985eebe7780989b0f9e89bfd
GitOrigin-RevId: 7ac4eefe4a6ded77189f4fa91d85f7198bb48355