Use robot-service@ as identity for sim robots
This makes sure the simulated robot is no longer dependent on the GCE
default service account, so it doesn't have more privilege than
necessary.

It also aligns with best practice:
https://cloud.google.com/compute/docs/access/service-accounts#scopes_best_practice
and eliminates some race conditions when restarting the metadata-server
on robot VMs.

I considered using Workload Identity when simulating a robot with GKE,
but we would need to annotate the service accounts on the robot
(potentially including service accounts inside app-* namespaces) but
this would not be useful outside of GKE.

See b/175282543.

Change-Id: Ibf64fce0d2b1c08eb394476afb11fe81d06fc40c
GitOrigin-RevId: 76048f0df10743829b016bb2c2fb1dae4e702662