Increase timeout in apps test
I'm not sure why, but it seems it takes sometimes ~one minute to delete
the old job, even though the underlying pod just runs `false` and should
terminate immediately. After bumping the timeout, the test passed 25
times in a row, whereas before it failed on the second attempt.

Fixes b/223331353.

Change-Id: Ice1760a9b7d51f47a3d0e052b62f2aa878c79618
GitOrigin-RevId: 86eb62aee3b09498c85d3cb13fae599c80bed02b