Shave ~40Mb from chart-assignment controller image.
Instead of running "helm init" through an init container we run it at build
time and include the config-tree as a container layer. This also avoid
having to run an init-container.

As one difference this does not contain a repository index (which we are
not actively using anyway).

I had to set the HELM_HOME var, since I was not able to forge a container
layer that would preserve the '.' in the helm dir.

Change-Id: Iba2233072cb6fba86314594e66fb3898689334bd
GitOrigin-RevId: c197ac04900fc4d70665ec833613fb330f0e0371