relay-server: shut down cleanly on SIGTERM
This aims to let ongoing requests complete before terminating the
server. To do this, we have to make sure that the long-running GET
/server/request?server=foo requests are stopped when the signal is
received, otherwise the server will take a long time to shut down
cleanly and the relay-clients will be stuck during this time.

If it still takes a long time to stop (for example, because it is
deadlocked or because there is a long running request like `kubectl
logs`) panic() - this should help debugging, especially with the
followup that enhances the health check.

Change-Id: Iec71091bd6f82ef3b6adb39d42e2f62aea41784a
GitOrigin-RevId: 3d6d5137ac2bb6ea4fdf977007d0c40a72359b2e