relay-server: add a health check for the broker
This aims to catch deadlocks in the broker and restart the server. It
could plausibly lead to undesired restarts when the server is loaded
(although the 10s timeout on the check should minimize this) but the
improved restart behavior in the parent commit should hopefully make
that less of a problem.

I hope this will help debug b/229946263 if it happens again.

Change-Id: Ibc18740117bd2b9141570d866aa69e77b0902bc9
GitOrigin-RevId: 0e319098bfbf27d30f50c59660960d0e672d019c