Tweak the prometheus chart to use node names instead of IPs.
Use a relabeling config to use the node-name instead of the IP on the
ServiceMonitor. Unfortunately there seem to be no way to setup a default.

Make some adjustments for k8s>1.18 that we use everywhere (avoids unavailable
scrape target).

Finally add a "cluster" label in addition to the robot label on the federation
monitor. Some grafana dashboards group by that label already.

Change-Id: I202c1edf4a705433d2df3b0fc33fccbb167e3dc4
GitOrigin-RevId: 6a099810689415e6cb4a691533e21284dc3758d7