Improve the integration test.
If one runs the test right after setting up a new cluster, it might fail
due to setup related restarts. The implemented logic is quite questionable,
since those restarts won't go way, but even if the remain consstant the
test fails.

The new logic in this change uses a map to track the restart count. If any
container has a restartcount > 0 we add it to the map and do a another
round. If the restart count remains unchanged we assum things have settled.

Note, this does not take exponential backoff into account (the test loop
has a fixed sleep of 10).

Change-Id: I701860ab283e2c639034305b747ae703d7fb58d4
GitOrigin-RevId: c82f5953435236050817342859bc20404868d583