Allow registering robots with human-acl@ tokens
If you have access to multiple cloud projects, the token printed by
gcloud can be used to access these projects. To avoid storing such a
sensitive token on a remote system where it could be compromised, this
change adds the option to generate a less privileged token, which is
still able to register robots by talking to GCS, GKE, and the token
vendor.

Please double-check my belief that this is safe because:

- Users need to be able to grant themselves serviceAccountTokenCreator
  on human-acl@ before they can create these tokens.
- Nothing else is acting as human-acl@ or creating tokens for it, so the
  new bindings are only usable as described above.

b/238401496

Change-Id: Ib23ce9b244dda0e7495e56a1bafac252816d0ad1
GitOrigin-RevId: 4085ec7a65df04cec10af6891864ec250af9154b