Add a --force_http2 option to the relay client.
It turns out that grpc might need the explicit use of go's http2 module.
Unfortunately we can't use that by default (unless --disable_http2 is given),
since it is causing 503 errors for e.g. http rest or html relaying.

To make this explicit, we add yet another flag.

Minor changes to debug logging (e.g. don't dump the http body, its not useful
for grpc and it modifies the instance).

Change-Id: I98d23dbf9be8e9bad1e8e5b0a842366c79fda394
GitOrigin-RevId: 6c3568aaa3f1e6075085b02c10dc51698d434ef9