metadata-server: store RobotAuth inside TokenHandler
This is a refactoring before the actual functionality change.
As we want to create token with an expire time of at least 67s we will need
to update the TokenSource. Therefore a RobotAuth is needed.
To avoid the need to create a new RobotAuth during ServeHTTP we store it inside
the TokenHandler. With this change TokenHandler has everything needed to create a
MetadataHandler so TokenHandler also got a factory function for
MetadataHandler.

Change-Id: Ia8abebbccb54c029d4ec241b8cf927c22603db48
GitOrigin-RevId: 912699b73de74cf20ac9d79917779179536d072e