Add publish key implementation for Cloud IoT repository.
A call to publishKey either updates the first public key of a device if the device exists already or creates a new device if the device identifier does not exist. This is less tested then it probably should be. A test for the API + IoT backend will be added in the next CL. Let me know if this should be tested better.

Change-Id: If71e8f46cb1736e9fbd7dee6599131c68054fc55
GitOrigin-RevId: 02abcf854a68cd470f9ae16b4d22bc65f97e18d6