Add verify token API endpoint logic.
This CL adds request validation and parameter parsing logic to the
verify token endpoint. The token can be supplied via header or URL
parameter in conformance with the Java-based token vendor. The parameter
alias "robot" for the parameter "robots" is not implemented. No internal
use for "robot" could be found.

b/175300687

Change-Id: I5950044ea0b4136c59973dcac03d9138154e00e8
GitOrigin-RevId: bb43ca61d1b87cd24401865da1df8f73e712bf1f