Add logic for verification for GCP IAM tokens.
This CL adds logic to verify if a given token has actAs permission
on a given GCP service account. This logic is connected to the verify
token endpoint. This CL does not include yet any caching logic to
unload the IAM backend (something the Java token vendor does). Testing
is done on the API layer with fake GCP responses.

b/175300687

Change-Id: Id427a65b66fbcf4330c948db6842340c2acf9bb7
GitOrigin-RevId: 2c2e17e87aab9a267a5e0b96b0abf46802458936