chart-assignment-controller: remove ClusterName verification
changing the cluster name inside the manifests for the cluster before applying them defeats the original purpose of adding the cluster name. And as the cluster name is no longer used to distinguish between chart assignments for the cloud and those for robots there seems to be no code left relying on ClusterName.

Change-Id: I376d3f62165952df6a5b3b3c1d1f39603b982b73
GitOrigin-RevId: da12bd86fbdb71cb4f6304cf6181a94188ed3ec9