replace deprecated/deleted api calls
replace calls to validation.ValidateClusterName and CachedDiscoveryClient.ServerResources with the direct
implementation as of k8s 1.23.
ValidateClusterName was just a variable with content of NameIsDNS1035Label.
ServerResources had a deprecation notice with advice to use ServerGroupsAndResources. ServerGroupsAndResources just has an addititional return value.

the calls got removed in later versions.

Change-Id: I395f5156221d33872617a2cad226fe37c6a02b59
GitOrigin-RevId: 5cf6b487afe30d4942c139d409f2c0bb6e94cae7