Increase selfsigned-ca duration to 1 year
The certificate for chart-assignment-controller is defaulted to 90 days
Due to how renewals work in cert-manager it can renew on a CA
which has a notBefore shorter than the cert.
cert-manager fails to renew the certificate invalidated by outdated CA.

With a CA lifetime longer than 3x the cert's lifetime, this is avoided
since the CA is renewed after 2/3rds of its lifetime.
Since the overlap time is longer than the certificate lifetime,
it is guaranteed to be renewed in this period.

BUG=b/259381824
TEST=none, this needs 30 days. See bug for reasoning

Change-Id: I425d4c4789902a21d54db5e030828a465541c1ab
GitOrigin-RevId: 5a6cffaf441cfae2fadd41b1a9185e0dab3a4cab