Stop controllers on SIGTERM
Our healthz server wasn't terminating on SIGTERM. Instead, we can use
the healthz server from controller-runtime, which has logic to handle
this. This means that the chart-assignment-controller will terminate on
SIGTERM, which should make updates and reboots faster.

Tested by deleting the pods and observing that they no longer spend 30s
in Terminating.

Change-Id: If26aeda5b1d99477cccedd7a7a30361abf961718
GitOrigin-RevId: e297157edeb689af169e68c09ce369a2283305a9