Use emptyDir for prometheus on robots
Set the storageSpec to use emptyDir on robots, so we can enable it
without writing emmc or other low-cycle storage to death.

We limit the retention set size to 200MiB.
This allows us to have a smaller allocation for the volume.

Size picked after experimentation, this should be around 2 days.

Change-Id: I8dcbc71202b0916bb92c94ae0631730e9715d377
GitOrigin-RevId: fc272dabbc93d55da6f382126b4d5744472add06