Fix Abseil parsing
There was quite a bit of confusion why the custom Abseil log parser
added in
https://github.com/googlecloudrobotics/core/commit/db089277d6e8066b489092abadbb1a3500ce0f47
didn't make it to deployments.

The reason is that this change was added to the values file, but
not to the actual YAML file
cloud-robotics/src/app_charts/base/robot/fluent-bit.yaml
which configures the deployment.

This change fixes it by doing the following:

1. Pin the chart template version in the script.
2. Reflect the following changes to the values file (they should've been
   applied to the values file in the first place):
    - https://github.com/googlecloudrobotics/core/commit/a4268b82ffb9289bf55d3bb6e9b5f9669abea1c5
    - https://github.com/googlecloudrobotics/core/commit/ad15bfb84e86cadd1bd4244f3bc8ba0b92ea656a
    - https://github.com/googlecloudrobotics/core/commit/c47099fae8f00dfc25951fe3b5f9e30eadf3ddf1
3. Extend the script to put a notice to the generated file.
4. Add a README explaining the change process.
5. Run the script.

Change-Id: Iace4ad15897fd01097beb360ad0da21399ded414
GitOrigin-RevId: ac3e737c512f0324265df08d348fbb4c38fc812e