Add CI/CD based on GitHub actions (#102)
This is copying the CI/CD setup from insrc with minimal changes
(adapting some paths, using a different cloud project for RBE).

I had to copy the `deployments/` folders for robco-integration-test and
robco-navtest as they are needed by the CI. However, I removed the Oauth
secrets from config.sh, which don't seem needed by the test. My plan for
now is to keep the duplicate setup, which I expect to change very
rarely.