Also patch the default/default SA.
A change to the order in which we update the GCR credentials introduced
a bug that prevented us from running pods with images from a private GCR
in the default namespace, which is triggered when setting up a new
cluster pointed at a CRC dev project.

The BUILD.bazel change was required by the pre-commit hook.