cr-syncer: Preserve observedGeneration semantics
The status's observedGeneration field lets a CR creator identify when a
controller has updated the status to reflect a change in the spec.
However, because it's tied to the generation field, which is managed by
the apiserver, it can't be directly copied between two clusters.
Instead, we can preserve the difference between generation and
observedGeneration, allowing the CR creator to wait until
generation==observedGeneration before looking at the status.