Get base images for go builds independent of rules_docker version (#136)
Closes issue #130.

This PR alters the bazel build to use `rules_oci` to pull go base images
using a version that's decoupled for the `rules_docker` version. This
allows for us to update the base image version without having to depend
on `rules_docker`.

The latest version of both the `base` and `static` distroless version
was used to avoid unexpected breakages.

Note that `oci_rules` `pull` supports using the `latest` tag. The macro
that pulls base images can be configured to always pull the latest
images and be up to date if we are okay with assuming the potential
upstream breakage risks it introduces.

---------

Signed-off-by: Alejo Carballude <alejocarballude@gmail.com>