Update ingress-nginx to 1.8.0 (#156)
This pulls in an Alpine update which should fix some CVE warnings. The
newer release is incompatible with k8s 1.23 but GKE has auto-updated to
1.24 everywhere now 1.23 is EOL.

I changed the RBAC policy based on the v1.3.0 release notes (needs
permission for coordination.k8s.io/leases). Don't ask me why upstream
added it to both the Role and the ClusterRole.

https://github.com/kubernetes/ingress-nginx/releases/tag/controller-v1.3.0

Tested by deploying to robco-rodrigoq and checking the nginx dashboards
in Grafana.