Fix handling of trailers in relay when chunking (#218)
The relay client splits responses from the backend into multiple chunks
based on MaxChunkSize and sends these chunks to the relay server.
Currently, the relay client includes HTTP trailers in the last chunk
sent - which makes sense - but the relay server code does not correctly
handle the case with more than one chunk and always returns the trailers
sent with the first chunk (ignoring the trailers in all other chunks).
Thus when chunking responses we currently do not return any trailers.

This PR fixes this issue in the relay server by handling trailers
similar to the message body and collecting all trailers from all chunks
before appending them to the HTTP response sent to the user client. We
add an integration test and unit tests that cover chunked responses.

This fixes the issue that relaying gRPC requests with responses larger
than MaxChunkSize leads to a `UNKNOWN` (`No status received`) error.
This also removes an existing
[TODO](https://github.com/googlecloudrobotics/core/blob/63b665ba7a0a9d9e46d0e387397ceade6db6659e/src/go/cmd/http-relay-server/server/server.go#L163)
which has spotted exactly this issue before.