fix leaking connections when user client closes connection (#222)
Same as #220, but ported to this repo.

---

When the user client closes the connection, it is not propagated to the
backend. For simple "one shot" requests, there is no huge impact as the
response would just be dropped, however for long running requests
(chunked encoding or websockets), the backend doesn't know that the user
client has closed the connection and it keeps sending new updates to the
relay client, which also doesn't know and so keep forwarding that to the
relay server.

The fix applied is:

* Add `StopRelayRequest` to the broker to "forget" a relaying request.
This will cause the relay server to response with a permanent error the
next time the relay client tries to send a response for that id. This
will in turn cause the relay client to close the connection to the
backend.
* There is a bug in the relay client when checking for backoff permanent
errors. When the backoff operation encounters a permanent error, it
actually unwraps it and return the underlying error, but the client is
still checking for `backoff.Permanent`, resulting in the check to always
fail and never handling it.

---------

Signed-off-by: Teo Koon Peng <koonpeng@google.com>