Hide integration tests from Buildkite CI
The Bazel team hosts a CI project on Buildkite, which expects to be able
to run all our tests with recent versions of Bazel. We like this because
the Bazel team file issues on our repo to tell us when a future Bazel
release will break our build, but to keep this, we need to make sure our
project is always green on their CI. I don't know why, but we just got a
report that our integration tests are red there due to missing
dependencies (kubectl, docker, credentials): #4380.

We can work around this by using the "manual" tag to exclude our tests
from target expansion, and then manually listing them in our own CI.

I will need to manually check that the tests are still running in
presubmit and postsubmit.