Delete obsolete CR_SYNCER_RBAC flag (#273)
This has been in use for a while. However, CR_SYNCER_RBAC is not
consistently true across all our projects - leading to excessive
robot-service@ privileges in some case. Deleting this simplifies the
code and improves security.

After this is applied, robot-service@ will be limited to "Kubernetes
Engine Cluster Viewer" in all other projects, which allows it to list
GKE clusters (via IAM) and other Robot CRs and ChartAssignments (via
RBAC) but not much more. This role is not to be confused with
"Kubernetes Engine Viewer" which would allow reading all resources in
the GKE clusters, rather than just listing clusters and reading select
resources.