Downgrade "empty response" error in discovery client to info log for metrics (#283)
This follows the precedent set in client-go's
[memcache](https://github.com/kubernetes/client-go/commit/0bc91705fa50a531b8a1ee9dca3ef063ca71bb0c).

The discovery client's `ServerGroupsAndResources()` call sometimes
throws up empty response errors if a metrics server (namely GCP's custom
metric server) neglects to populate its resource list. This is not
considered an error.

The [bump of client-go's
version](https://github.com/googlecloudrobotics/core/pull/264) fixed
`memcache.go`, but we also need to account for it in our code.

Tested with a test deploy, the CAC works as expected.

---

**Note**: Logging the downgraded error generates quite a few spurious
logs on startup (in fact, I think it generates one log PER chart update
call, so each time a chart is updated, we'll get a printout...) Might be
scary to a user, so we might consider just not printing, throttling the
print, or explicitly writing "(NOT AN ERROR)" in the log. What do y'all
think?

I am quite for just ignoring the error. So much so I'm going to push
that change now.


![image](https://github.com/googlecloudrobotics/core/assets/20265670/edc73899-ca9b-4ec4-8fc2-b5a672b85c4c)

Signed-off-by: methylDragon <methylDragon@google.com>