Correctly apply relay config (#284)
https://github.com/googlecloudrobotics/core/pull/260 had no effect
because the default config settings require additional support in
main.go to make it through the client.

This eliminates the mistake and reduces the number of lines needed per
config setting, especially for those that don't (yet?) need a flag like
BackendResponseTimeout or ReadIdleTimeout.

This should actually fix issues we encountered when (re)configuring the
network
on the nodes. b/313612734

All credit to Koon Peng for figuring out my mistake here!
b/316473267#comment15. We should think about banning changes that are
authored and reviewed only by managers as we obviously don't have the
attention span to catch issues like this.