Simplify rbe_config (#295)
We realized the only thing used from the rbe_config directory is a
single `platform` rule in `bazel/rbe_config/config/BUILD`. This is
because the full hermetic toolchain is specified through WORKSPACE, we
do not use the https://github.com/bazelbuild/bazel-toolchains auto
generated one.

Move this rule to the parent BUILD and rename it to be more explicit.
Delete the rest of the directory.

Also adds `.bazelversion` to make this repo bazelisk compatible.