Remove min_master_version
Because we hardcoded this, we have to continuously bump it. I'm not sure
why we ever added it but I don't think we need it now, as GKE only
supports recent versions and the Terraform docs say this is optional.

Terraform can handle this change with no problem:

```
google_container_cluster.cloud-robotics will be updated in-place
```

Fixes #316.