Extend the backup script to also backup the pubkeys. (#318)
This is mostly FYI. I don't like the hacks it requires. If I run this
against g-w I get a 0.5 mb yaml and there are a lot of pubkeys:
```
egrep "^kind:" /tmp/crc-backup.yaml | sort | uniq -c
    706 kind: ConfigMap
     78 kind: Robot
```
Some of them are stale, but token-vendor can't do the cleanup, because
the robot-cr is not necessarily available in the cloud.

Also the filtering is sub-par. WDYT if I change token-vendor to label
them (and backfill the label via script).