Deploy navtest on cloud build (#325)
Because of Terraform being too old (b/323509860), I'm blocked on landing
https://github.com/googlecloudrobotics/core/pull/323 since
`deploy_navtest.sh` does not work with WIF.

This circumvents the problem by running it on Cloud Build (using "within
cloud credentials") instead of directly on the action runner.

I had to remove curl `--fail-with-body` because the version of curl in
the cloud build image is too old and it is non-trivial to upgrade it on
Ubuntu 20. And building a custom image just for this seemed overkill. In
https://github.com/googlecloudrobotics/core/pull/239, the motivation was
not this particular call site but the GitHub API calls, which are still
kept as is.