chart-assignment-controller: use the same resource builder as synk (#338)
see b/207673353
there was a difference between manifest parsing in synk and in the
chart-assignment-controller. the handling of 'kind: List' was causing an
infinite resource creation by the controller but was handled correctly
by synk. with this change manifest parsing is done by
cli-runtime/pkg/resource as is already the case for synk.