Remove request_path from broker metrics (#340)
This was previously useful for tracking how the relay was used, but
there's a risk that if the URLs change frequently, the cardinality
increases without bound. I'm not aware that anyone uses this, and can't
find any references to request_path in any codebase, so it seems like
high time to remove this Prometheus outage waiting to happen.

b/294533220