Make integration test use GKE robot (#348)
This fixes https://github.com/googlecloudrobotics/core/issues/346 and
fixes https://github.com/googlecloudrobotics/core/issues/319. It removes
`install_k8s_on_robot.sh` and `setup-robot-sim.sh`, which are now
deprecated.

We can run both the relay_test and the go_default_test against a "GKE
robot" instead of one running in a VM. This removes a lot of complexity
from the testing infrastructure. The downside is that we lose some test
coverage of the onprem metadata server, but that seems like a price we
have to pay.

Another advantage is that with this change, `integration_test.sh` can
run locally and not only in the Cloud Build environment, which is useful
for debugging.

Green run:
https://github.com/googlecloudrobotics/core/actions/runs/8233748802/job/22513900247

Follow-ups: Remove robot-sim TF. Remove TF around custom cloud build
pools, which were only needed for ssh to GCE.