Only consider scheduled postsubmit runs
Please review carefully as I'm not confident I'm properly closing the
hole here.

Without this parameter, the API will also return commits from PRs, which
may not have been reviewed and could be malicious.

This has the downside that if you're trying to fix the release workflow
you need to change the schedules or be limited to one attempted fix per
day, but I'd rather close the hole before looking for a better solution.

I was unable to identify a way to limit actions in general to the main
branch, and wasn't immediately sure how to do that check on the result
of the API call (is there a better way than `git log main | grep
$green_sha`?) but it seems like limiting to `event=schedule` should be
sufficient.

http://b/348316770