Allow forward authentication with robot JWT (#409)
This allows users to configure proxies (e.g. nginx) to defer
authentication to the token vendor not just via the OIDC token, but
already with the JWT. Providing this method allows to authenticate "from
the first request" without having to first allow traffict to the token
vendor from the client. Which is useful to allow an authenticated proxy
that's on the route between the client and the token-vendor.

TESTED=auto, not yet in practice