Actually drop container_network metrics
We added container_network.* to the wrong part of the values so they are
still the highest-cardinality metrics (around 15% of all metrics).

Tested by using `bazel build
src/app_charts/prometheus/prometheus-operator-chart.robot.yaml` to build
the YAML and diffing:

```
--- /tmp/before.yaml    2024-10-10 15:40:11.686509000 +0000
+++ /tmp/after.yaml     2024-10-10 15:52:02.595166797 +0000
@@ -1192,7 +1192,7 @@
       sourceLabels:
       - __name__
     - action: drop
-      regex: container_spec.*
+      regex: container_spec.*|container_network.*
       sourceLabels:
       - __name__
     - action: drop
```

then manually applying this change with k9s to a lab system and
verifying that the metric stops flowing to the cloud.