Add smartctl-exporter
This exports metrics for both SATA and NVMe drives (even if it
annoyingly generates a completely different set of metrics in the two
cases). The container is relatively light at 30MB (!).

This will let us see historic drive temperatures and error rates if we
start to notice drive-related problems in future.

I considered putting this behind a Helm value, but I expect we want
these metrics everywhere, and if they break something we can just roll
back the CRC change.

Tested with kubectl apply on the YAML file.