Fix IDE integration with `rules_go` (#463)
IDE integration no longer works when following the guide
https://github.com/bazel-contrib/rules_go/wiki/Editor-setup.

The cause of the issue is that because we have
`--incompatible_use_plus_in_repo_names`, `rules_go` tries to query
targets like `@@rules_go+//:stdlib`. But because we only enable that
flag for build commands, the query command is still expecting
`@@rules_go~//:stdlib`. I also need to update bazel to 7.4.1 for it to
work.

To test: `echo '{}' | bazel run --
@io_bazel_rules_go//go/tools/gopackagesdriver ./... builtin > /dev/null`

Signed-off-by: Teo Koon Peng <koonpeng@google.com>