Cleanup robotauth tokensource config.
Leave the scopes empty
This is optional and we already sent the data else where. I've deployed this
to robco-ensonic and connected minikube to it. I also started hello-k8s and
observed that the logs are captured in the cloud.

Add some more comments on the other fields as well.