Replace logName field to avoid a high cardinality field on stackdriver
When aggregating logs from many peers, we'll see a lot of variance in this
field. All teh infomation is already conatin in teh record, so replace the
field with a more static valud. This is a similar approach that GKE takes.

Note: the stackdriver plugin adds the projects/<project-id>/logs prefix.

See https://github.com/fluent/fluent-bit/issues/9897