Implement cr-syncer-auth-webhook
This adds code to verify and decode the JWTs and then to only allow
requests for resources corresponding to this robot.

Because this doesn't talk to the GKE apiserver itself, it is very
limited:

- it hardcodes the list of supported CRs
- it assumes that any CR with the suffix "robotName" is visible to that
  robot (could be exploited by creating robots called 0, 1, 2, ...)

I've tested this by setting --use-robot-jwt=true on the cr-syncer on a
test system, but the new codepath won't be used until that becomes the
default.