Remove support for spec-source=robot
This is unused and makes it harder to add support for robot JWTs.
Previously we printed an INFO log for all non-synced CRs in the system
(eg istio, prometheus, cert-manager CRs) which is a bit noisy. To make
the error more obvious if someone uses spec-source=robot, I've silenced
the messages when spec-source is "" and made them errors for other
values.