cr-syncer: list robottypes in Health check (#512)
For simplicity, the health check doesn't limit its request to a single
resource. This generates a "robot impersonation rejected" log message in
the cloud when it tries to list all Robot CRs. This doesn't break the
check, since it only fails on 401 but this gets a 403. However, the log
message is confusing.

Instead, we can check for robottypes, which allows a successful &
simpler request.

Tested by pushing the cr-syncer change to the GKE cluster created by
scripts/robot-sim.sh (with --use-robot-jwt=true) and verifying that the
error logs from the cr-syncer-auth-webhook stop.