Allow switching off the cr-syncer-auth-webhook (#536)
We previously mistakenly assumed .Values.onprem_federation was a
boolean, but it seems that deploy.sh passes all values through as
strings. I don't remember if there is a good reason for this...

To align with other switches like .Values.app_management, let's
stringly-type this too.

Tested by setting ONPREM_FEDERATION=true/false in config.sh, deploying,
and checking for the deployment in GKE.