Restrict maintenance window to weekend (#575)
We observed a prod maintenance event on 20 October 2025 at 08:49:03
GMT+2.

My hypothesis is that because we have a >24h window (05:00 on 2025-04-05
to 19:00 on 2025-04-*06*) starting on Saturday *and* Sunday, we end up
with two overlapping windows, or an overall window that lasts until
Monday 19:00 UTC. I don't know how to test but I hope that removing
`,SU` will prevent this.