http-relay-server: Add a flag to set inactive request timeout (#597)
Please review the individual commits to separate the refactor from the
new flag:

- refactor: use Config struct for server.go
- feat: add flag to change request timeout

We've seen outages where a critical request goes from taking slightly
less than 60s to a bit more, and it's hard to response to these as the
timeout is hard-coded. This should give us the flexibility to change it.